bugfix:mtp embedding cache lifecycle.

+++ xllm/core/framework/batch/batch_input_builder.h
@@ -100,7 +100,7 @@ class BatchInputBuilder {
     std::vector<float> acc_logprob_vec;
 
     // Additional data
-    std::vector<int32_t> embedding_ids;
+    std::vector<int64_t> embedding_ids;
     std::vector<std::string> request_ids;
     std::vector<int32_t> extra_token_ids;
     std::vector<TransferKVInfo> transfer_kv_infos;
